moved apply cart price rules to CartPriceCalculator (not in Cart any more) should prevent endless loops when cart is loaded within pricing rule actions
